+++ consumer/consumererror/combine.go
@@ -24,8 +24,8 @@ import (
 // If any of the errors in errs are Permanent then the returned
 // error will also be Permanent.
 //
-// Any signal data associated with a PartialError will be discarded
-// and the resulting error will not be a PartialError.
+// Any signal data associated with an error from this package
+// will be discarded.
 func Combine(errs []error) error {
 	numErrors := len(errs)
 	if numErrors == 0 {