+++ exporter/exporterhelper/metricshelper.go
@@ -46,8 +46,8 @@ func newMetricsRequest(ctx context.Context, md pdata.Metrics, pusher PushMetrics
 }
 
 func (req *metricsRequest) onError(err error) request {
-	metricsError := &consumererror.Metrics{}
-	if consumererror.AsMetrics(err, metricsError) {
+	var metricsError consumererror.Metrics
+	if consumererror.AsMetrics(err, &metricsError) {
 		return newMetricsRequest(req.ctx, metricsError.GetMetrics(), req.pusher)
 	}
 	return req